+++ squeal-postgresql/squeal-postgresql.cabal
@@ -33,7 +33,6 @@ library
     Squeal.PostgreSQL.Definition.View
     Squeal.PostgreSQL.Expression
     Squeal.PostgreSQL.Expression.Aggregate
-    Squeal.PostgreSQL.Expression.Aggregate2
     Squeal.PostgreSQL.Expression.Array
     Squeal.PostgreSQL.Expression.Comparison
     Squeal.PostgreSQL.Expression.Composite
@@ -47,14 +46,12 @@ library
     Squeal.PostgreSQL.Expression.Range
     Squeal.PostgreSQL.Expression.Set
     Squeal.PostgreSQL.Expression.Sort
-    Squeal.PostgreSQL.Expression.Sort2
     Squeal.PostgreSQL.Expression.Subquery
     Squeal.PostgreSQL.Expression.Text
     Squeal.PostgreSQL.Expression.TextSearch
     Squeal.PostgreSQL.Expression.Time
     Squeal.PostgreSQL.Expression.Type
     Squeal.PostgreSQL.Expression.Window
-    Squeal.PostgreSQL.Expression.Window2
     Squeal.PostgreSQL.List
     Squeal.PostgreSQL.Manipulation
     Squeal.PostgreSQL.PG
