+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Aggregate.hs
@@ -17,7 +17,9 @@ Aggregate functions
   , LambdaCase
   , MultiParamTypeClasses
   , OverloadedStrings
+  , PatternSynonyms
   , PolyKinds
+  , StandaloneDeriving
   , TypeFamilies
   , TypeOperators
   , UndecidableInstances
@@ -26,23 +28,26 @@ Aggregate functions
 module Squeal.PostgreSQL.Expression.Aggregate
   ( -- * Aggregate
     Aggregate (..)
-    -- * Distinction
-  , Distinction (..)
+  , AggregateArg (..)
+  , pattern All
+  , pattern Alls
+  , pattern Distinct
+  , pattern Distincts
+  , FilterWhere (..)
     -- * Aggregate Types
   , PGSum
   , PGAvg
   ) where
 
-import Control.DeepSeq
 import Data.ByteString (ByteString)
-import Data.Kind
 import GHC.TypeLits
 
-import qualified GHC.Generics as GHC
 import qualified Generics.SOP as SOP
 
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Expression
+import Squeal.PostgreSQL.Expression.Logic
+import Squeal.PostgreSQL.Expression.Sort
 import Squeal.PostgreSQL.List
 import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Schema
@@ -55,8 +60,7 @@ import Squeal.PostgreSQL.Schema
 `Aggregate` functions can be used as `Grouped` `Expression`s as well
 as `Squeal.PostgreSQL.Expression.Window.WindowFunction`s.
 -}
-class Aggregate expr1 exprN aggr
-  | aggr -> expr1, aggr -> exprN where
+class Aggregate arg expr | expr -> arg where
 
   -- | A special aggregation that does not require an input
   --
@@ -67,7 +71,7 @@ class Aggregate expr1 exprN aggr
   -- in printSQL expression
   -- :}
   -- count(*)
-  countStar :: aggr ('NotNull 'PGint8)
+  countStar :: expr lat with db params from ('NotNull 'PGint8)
 
   -- | >>> :{
   -- let
@@ -77,8 +81,8 @@ class Aggregate expr1 exprN aggr
   -- :}
   -- count(ALL "col")
   count
-    :: expr1 ty
-    -> aggr ('NotNull 'PGint8)
+    :: arg '[ty] lat with db params from
+    -> expr lat with db params from ('NotNull 'PGint8)
 
   -- | >>> :{
   -- let
@@ -88,23 +92,23 @@ class Aggregate expr1 exprN aggr
   -- :}
   -- sum(DISTINCT "col")
   sum_
-    :: expr1 (null ty)
-    -> aggr ('Null (PGSum ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGSum ty))
 
   -- | input values, including nulls, concatenated into an array
   arrayAgg
-    :: expr1 ty
-    -> aggr ('Null ('PGvararray ty))
+    :: arg '[ty] lat with db params from
+    -> expr lat with db params from ('Null ('PGvararray ty))
 
   -- | aggregates values as a JSON array
   jsonAgg
-    :: expr1 ty
-    -> aggr ('Null 'PGjson)
+    :: arg '[ty] lat with db params from
+    -> expr lat with db params from ('Null 'PGjson)
 
   -- | aggregates values as a JSON array
   jsonbAgg
-    :: expr1 ty
-    -> aggr ('Null 'PGjsonb)
+    :: arg '[ty] lat with db params from
+    -> expr lat with db params from ('Null 'PGjsonb)
 
   {- |
   the bitwise AND of all non-null input values, or null if none
@@ -119,8 +123,8 @@ class Aggregate expr1 exprN aggr
   -}
   bitAnd
     :: int `In` PGIntegral
-    => expr1 (null int)
-    -> aggr ('Null int)
+    => arg '[null int] lat with db params from
+    -> expr lat with db params from ('Null int)
 
   {- |
   the bitwise OR of all non-null input values, or null if none
@@ -135,8 +139,8 @@ class Aggregate expr1 exprN aggr
   -}
   bitOr
     :: int `In` PGIntegral
-    => expr1 (null int)
-    -> aggr ('Null int)
+    => arg '[null int] lat with db params from
+    -> expr lat with db params from ('Null int)
 
   {- |
   true if all input values are true, otherwise false
@@ -144,14 +148,14 @@ class Aggregate expr1 exprN aggr
   >>> :{
   let
     winFun :: WindowFunction  'Ungrouped '[] with db params '[tab ::: '["col" ::: null 'PGbool]] ('Null 'PGbool)
-    winFun = boolAnd #col
+    winFun = boolAnd (Window #col)
   in printSQL winFun
   :}
   bool_and("col")
   -}
   boolAnd
-    :: expr1 (null 'PGbool)
-    -> aggr ('Null 'PGbool)
+    :: arg '[null 'PGbool] lat with db params from
+    -> expr lat with db params from ('Null 'PGbool)
 
   {- |
   true if at least one input value is true, otherwise false
@@ -165,8 +169,8 @@ class Aggregate expr1 exprN aggr
   bool_or(ALL "col")
   -}
   boolOr
-    :: expr1 (null 'PGbool)
-    -> aggr ('Null 'PGbool)
+    :: arg '[null 'PGbool] lat with db params from
+    -> expr lat with db params from ('Null 'PGbool)
 
   {- |
   equivalent to `boolAnd`
@@ -180,253 +184,296 @@ class Aggregate expr1 exprN aggr
   every(DISTINCT "col")
   -}
   every
-    :: expr1 (null 'PGbool)
-    -> aggr ('Null 'PGbool)
+    :: arg '[null 'PGbool] lat with db params from
+    -> expr lat with db params from ('Null 'PGbool)
 
   {- |maximum value of expression across all input values-}
   max_
-    :: expr1 (null ty)
-    -> aggr ('Null ty)
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null ty)
 
   -- | minimum value of expression across all input values
   min_
-    :: expr1 (null ty)
-    -> aggr ('Null ty)
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null ty)
 
   -- | the average (arithmetic mean) of all input values
   avg
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
   {- | correlation coefficient
 
   >>> :{
   let
     expression :: Expression ('Grouped g) '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGfloat8)
-    expression = corr (All (#y *: #x))
+    expression = corr (Alls (#y *: #x))
   in printSQL expression
   :}
   corr(ALL "y", "x")
   -}
   corr
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   {- | population covariance
 
   >>> :{
   let
     expression :: Expression ('Grouped g) '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGfloat8)
-    expression = covarPop (All (#y *: #x))
+    expression = covarPop (Alls (#y *: #x))
   in printSQL expression
   :}
   covar_pop(ALL "y", "x")
   -}
   covarPop
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   {- | sample covariance
 
   >>> :{
   let
     winFun :: WindowFunction  'Ungrouped '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGfloat8)
-    winFun = covarSamp (#y *: #x)
+    winFun = covarSamp (Windows (#y *: #x))
   in printSQL winFun
   :}
   covar_samp("y", "x")
   -}
   covarSamp
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   {- | average of the independent variable (sum(X)/N)
 
   >>> :{
   let
     expression :: Expression ('Grouped g) '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGfloat8)
-    expression = regrAvgX (All (#y *: #x))
+    expression = regrAvgX (Alls (#y *: #x))
   in printSQL expression
   :}
   regr_avgx(ALL "y", "x")
   -}
   regrAvgX
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   {- | average of the dependent variable (sum(Y)/N)
 
   >>> :{
   let
     winFun :: WindowFunction  'Ungrouped '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGfloat8)
-    winFun = regrAvgY (#y *: #x)
+    winFun = regrAvgY (Windows (#y *: #x))
   in printSQL winFun
   :}
   regr_avgy("y", "x")
   -}
   regrAvgY
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   {- | number of input rows in which both expressions are nonnull
 
   >>> :{
   let
     winFun :: WindowFunction  'Ungrouped '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGint8)
-    winFun = regrCount (#y *: #x)
+    winFun = regrCount (Windows (#y *: #x))
   in printSQL winFun
   :}
   regr_count("y", "x")
   -}
   regrCount
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGint8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGint8)
 
   {- | y-intercept of the least-squares-fit linear equation determined by the (X, Y) pairs
   >>> :{
   let
     expression :: Expression ('Grouped g) '[] c s p '[t ::: '["x" ::: 'NotNull 'PGfloat8, "y" ::: 'NotNull 'PGfloat8]] ('Null 'PGfloat8)
-    expression = regrIntercept (All (#y *: #x))
+    expression = regrIntercept (Alls (#y *: #x))
   in printSQL expression
   :}
   regr_intercept(ALL "y", "x")
   -}
   regrIntercept
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   -- | @regr_r2(Y, X)@, square of the correlation coefficient
   regrR2
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   -- | @regr_slope(Y, X)@, slope of the least-squares-fit linear equation
   -- determined by the (X, Y) pairs
   regrSlope
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   -- | @regr_sxx(Y, X)@, sum(X^2) - sum(X)^2/N
   -- (“sum of squares” of the independent variable)
   regrSxx
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   -- | @regr_sxy(Y, X)@, sum(X*Y) - sum(X) * sum(Y)/N
   -- (“sum of products” of independent times dependent variable)
   regrSxy
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   -- | @regr_syy(Y, X)@, sum(Y^2) - sum(Y)^2/N
   -- (“sum of squares” of the dependent variable)
   regrSyy
-    :: exprN '[null 'PGfloat8, null 'PGfloat8]
-    -> aggr ('Null 'PGfloat8)
+    :: arg '[null 'PGfloat8, null 'PGfloat8] lat with db params from
+    -> expr lat with db params from ('Null 'PGfloat8)
 
   -- | historical alias for `stddevSamp`
   stddev
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
   -- | population standard deviation of the input values
   stddevPop
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
   -- | sample standard deviation of the input values
   stddevSamp
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
   -- | historical alias for `varSamp`
   variance
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
   -- | population variance of the input values
   -- (square of the population standard deviation)
   varPop
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
   -- | sample variance of the input values
   -- (square of the sample standard deviation)
   varSamp
-    :: expr1 (null ty)
-    -> aggr ('Null (PGAvg ty))
+    :: arg '[null ty] lat with db params from
+    -> expr lat with db params from ('Null (PGAvg ty))
 
 {- |
-`Distinction`s are used for the input of `Aggregate` `Expression`s.
+`AggregateArg`s are used for the input of `Aggregate` `Expression`s.
 `All` invokes the aggregate once for each input row.
 `Distinct` invokes the aggregate once for each distinct value of the expression
-(or distinct set of values, for multiple expressions) found in the input
+(or distinct set of values, for multiple expressions) found in the input.
 -}
-data Distinction (expr :: kind -> Type) (ty :: kind)
-  = All (expr ty)
-  | Distinct (expr ty)
-  deriving (GHC.Generic,Show,Eq,Ord)
-instance NFData (Distinction (Expression grp lat with db params from) ty)
-instance RenderSQL (Distinction (Expression grp lat with db params from) ty) where
-  renderSQL = \case
-    All x -> "ALL" <+> renderSQL x
-    Distinct x -> "DISTINCT" <+> renderSQL x
-instance SOP.SListI tys => RenderSQL
-  (Distinction (NP (Expression grp lat with db params from)) tys) where
-    renderSQL = \case
-      All xs -> "ALL" <+> renderCommaSeparated renderSQL xs
-      Distinct xs -> "DISTINCT" <+> renderCommaSeparated renderSQL xs
+data AggregateArg
+  (xs :: [NullType])
+  (lat :: FromType)
+  (with :: FromType)
+  (db :: SchemasType)
+  (params :: [NullType])
+  (from :: FromType)
+  = AggregateAll
+  { aggregateArgs :: NP (Expression 'Ungrouped lat with db params from) xs
+  , aggregateOrder :: [SortExpression 'Ungrouped lat with db params from]
+  , aggregateFilter :: [Condition 'Ungrouped lat with db params from] }
+  | AggregateDistinct
+  { aggregateArgs :: NP (Expression 'Ungrouped lat with db params from) xs
+  , aggregateOrder :: [SortExpression 'Ungrouped lat with db params from]
+  , aggregateFilter :: [Condition 'Ungrouped lat with db params from] }
 
-instance Aggregate
-  (Distinction (Expression 'Ungrouped lat with db params from))
-  (Distinction (NP (Expression 'Ungrouped lat with db params from)))
-  (Expression ('Grouped bys) lat with db params from) where
-    countStar = UnsafeExpression "count(*)"
-    count = unsafeAggregate1 "count"
-    sum_ = unsafeAggregate1 "sum"
-    arrayAgg = unsafeAggregate1 "array_agg"
-    jsonAgg = unsafeAggregate1 "json_agg"
-    jsonbAgg = unsafeAggregate1 "jsonb_agg"
-    bitAnd = unsafeAggregate1 "bit_and"
-    bitOr = unsafeAggregate1 "bit_or"
-    boolAnd = unsafeAggregate1 "bool_and"
-    boolOr = unsafeAggregate1 "bool_or"
-    every = unsafeAggregate1 "every"
-    max_ = unsafeAggregate1 "max"
-    min_ = unsafeAggregate1 "min"
-    avg = unsafeAggregate1 "avg"
-    corr = unsafeAggregateN "corr"
-    covarPop = unsafeAggregateN "covar_pop"
-    covarSamp = unsafeAggregateN "covar_samp"
-    regrAvgX = unsafeAggregateN "regr_avgx"
-    regrAvgY = unsafeAggregateN "regr_avgy"
-    regrCount = unsafeAggregateN "regr_count"
-    regrIntercept = unsafeAggregateN "regr_intercept"
-    regrR2 = unsafeAggregateN "regr_r2"
-    regrSlope = unsafeAggregateN "regr_slope"
-    regrSxx = unsafeAggregateN "regr_sxx"
-    regrSxy = unsafeAggregateN "regr_sxy"
-    regrSyy = unsafeAggregateN "regr_syy"
-    stddev = unsafeAggregate1 "stddev"
-    stddevPop = unsafeAggregate1 "stddev_pop"
-    stddevSamp = unsafeAggregate1 "stddev_samp"
-    variance = unsafeAggregate1 "variance"
-    varPop = unsafeAggregate1 "var_pop"
-    varSamp = unsafeAggregate1 "var_samp"
+instance SOP.SListI xs => RenderSQL (AggregateArg xs lat with db params from) where
+  renderSQL = \case
+    AggregateAll args sorts filters ->
+      parenthesized ("ALL" <+> renderCommaSeparated renderSQL args)
+      <> renderSQL sorts
+      <> renderFilters filters
+    AggregateDistinct args sorts filters ->
+      parenthesized ("DISTINCT" <+> renderCommaSeparated renderSQL args)
+      <> renderSQL sorts
+      <> renderFilters filters
+    where
+      renderFilter wh = "FILTER" <+> parenthesized ("WHERE" <+> wh)
+      renderFilters = \case
+        [] -> ""
+        wh:whs -> " " <> renderFilter (renderSQL (foldr (.&&) wh whs))
+
+instance OrderBy (AggregateArg xs) 'Ungrouped where
+  orderBy sorts1 = \case
+    AggregateAll xs sorts0 whs -> AggregateAll xs (sorts0 ++ sorts1) whs
+    AggregateDistinct xs sorts0 whs -> AggregateDistinct xs (sorts0 ++ sorts1) whs
+
+pattern All
+  :: Expression 'Ungrouped lat with db params from x
+  -> AggregateArg '[x] lat with db params from
+pattern All x = Alls (x :* Nil)
+
+pattern Alls
+  :: NP (Expression 'Ungrouped lat with db params from) xs
+  -> AggregateArg xs lat with db params from
+pattern Alls xs = AggregateAll xs [] []
+
+pattern Distinct
+  :: Expression 'Ungrouped lat with db params from x
+  -> AggregateArg '[x] lat with db params from
+pattern Distinct x = Distincts (x :* Nil)
+
+pattern Distincts
+  :: NP (Expression 'Ungrouped lat with db params from) xs
+  -> AggregateArg xs lat with db params from
+pattern Distincts xs = AggregateDistinct xs [] []
+
+class FilterWhere arg grp where
+  filterWhere
+    :: Condition grp lat with db params from
+    -> arg xs lat with db params from
+    -> arg xs lat with db params from
+instance FilterWhere AggregateArg 'Ungrouped where
+  filterWhere wh = \case
+    AggregateAll xs sorts whs -> AggregateAll xs sorts (wh : whs)
+    AggregateDistinct xs sorts whs -> AggregateDistinct xs sorts (wh : whs)
+
+instance Aggregate AggregateArg (Expression ('Grouped bys)) where
+  countStar = UnsafeExpression "count(*)"
+  count = unsafeAggregate "count"
+  sum_ = unsafeAggregate "sum"
+  arrayAgg = unsafeAggregate "array_agg"
+  jsonAgg = unsafeAggregate "json_agg"
+  jsonbAgg = unsafeAggregate "jsonb_agg"
+  bitAnd = unsafeAggregate "bit_and"
+  bitOr = unsafeAggregate "bit_or"
+  boolAnd = unsafeAggregate "bool_and"
+  boolOr = unsafeAggregate "bool_or"
+  every = unsafeAggregate "every"
+  max_ = unsafeAggregate "max"
+  min_ = unsafeAggregate "min"
+  avg = unsafeAggregate "avg"
+  corr = unsafeAggregate "corr"
+  covarPop = unsafeAggregate "covar_pop"
+  covarSamp = unsafeAggregate "covar_samp"
+  regrAvgX = unsafeAggregate "regr_avgx"
+  regrAvgY = unsafeAggregate "regr_avgy"
+  regrCount = unsafeAggregate "regr_count"
+  regrIntercept = unsafeAggregate "regr_intercept"
+  regrR2 = unsafeAggregate "regr_r2"
+  regrSlope = unsafeAggregate "regr_slope"
+  regrSxx = unsafeAggregate "regr_sxx"
+  regrSxy = unsafeAggregate "regr_sxy"
+  regrSyy = unsafeAggregate "regr_syy"
+  stddev = unsafeAggregate "stddev"
+  stddevPop = unsafeAggregate "stddev_pop"
+  stddevSamp = unsafeAggregate "stddev_samp"
+  variance = unsafeAggregate "variance"
+  varPop = unsafeAggregate "var_pop"
+  varSamp = unsafeAggregate "var_samp"
 
 -- | escape hatch to define aggregate functions
-unsafeAggregate1
-  :: ByteString -- ^ aggregate function
-  -> Distinction (Expression 'Ungrouped lat with db params from) x
-  -> Expression ('Grouped bys) lat with db params from y
-unsafeAggregate1 fun x = UnsafeExpression $ fun <> parenthesized (renderSQL x)
-
-unsafeAggregateN
+unsafeAggregate
   :: SOP.SListI xs
   => ByteString -- ^ function
-  -> Distinction (NP (Expression 'Ungrouped lat with db params from)) xs
+  -> AggregateArg xs lat with db params from
   -> Expression ('Grouped bys) lat with db params from y
-unsafeAggregateN fun xs = UnsafeExpression $ fun <> parenthesized (renderSQL xs)
+unsafeAggregate fun xs = UnsafeExpression $ fun <> renderSQL xs
 
 -- | A type family that calculates `PGSum` `PGType` of
 -- a given argument `PGType`.
