+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Window.hs
@@ -19,6 +19,7 @@ Window functions and definitions
   , LambdaCase
   , MultiParamTypeClasses
   , OverloadedStrings
+  , PatternSynonyms
   , RankNTypes
   , KindSignatures
 #-}
@@ -30,6 +31,9 @@ module Squeal.PostgreSQL.Expression.Window
     -- * Window Function
     -- ** Types
   , WindowFunction (..)
+  , WindowArg (..)
+  , pattern Window
+  , pattern Windows
   , WinFun0
   , WinFun1
   , WinFunN
@@ -50,7 +54,7 @@ module Squeal.PostgreSQL.Expression.Window
   ) where
 
 import Control.DeepSeq
-import Data.ByteString
+import Data.ByteString (ByteString)
 
 import qualified GHC.Generics as GHC
 import qualified Generics.SOP as SOP
@@ -58,47 +62,45 @@ import qualified Generics.SOP as SOP
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Aggregate
+import Squeal.PostgreSQL.Expression.Logic
 import Squeal.PostgreSQL.Expression.Sort
 import Squeal.PostgreSQL.List
 import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Schema
 
-instance Aggregate
-  (Expression grp lat with db params from)
-  (NP (Expression grp lat with db params from))
-  (WindowFunction grp lat with db params from) where
-    countStar = UnsafeWindowFunction "count(*)"
-    count = unsafeWindowFunction1 "count"
-    sum_ = unsafeWindowFunction1 "sum"
-    arrayAgg = unsafeWindowFunction1 "array_agg"
-    jsonAgg = unsafeWindowFunction1 "json_agg"
-    jsonbAgg = unsafeWindowFunction1 "jsonb_agg"
-    bitAnd = unsafeWindowFunction1 "bit_and"
-    bitOr = unsafeWindowFunction1 "bit_or"
-    boolAnd = unsafeWindowFunction1 "bool_and"
-    boolOr = unsafeWindowFunction1 "bool_or"
-    every = unsafeWindowFunction1 "every"
-    max_ = unsafeWindowFunction1 "max"
-    min_ = unsafeWindowFunction1 "min"
-    avg = unsafeWindowFunction1 "avg"
-    corr = unsafeWindowFunctionN "corr"
-    covarPop = unsafeWindowFunctionN "covar_pop"
-    covarSamp = unsafeWindowFunctionN "covar_samp"
-    regrAvgX = unsafeWindowFunctionN "regr_avgx"
-    regrAvgY = unsafeWindowFunctionN "regr_avgy"
-    regrCount = unsafeWindowFunctionN "regr_count"
-    regrIntercept = unsafeWindowFunctionN "regr_intercept"
-    regrR2 = unsafeWindowFunctionN "regr_r2"
-    regrSlope = unsafeWindowFunctionN "regr_slope"
-    regrSxx = unsafeWindowFunctionN "regr_sxx"
-    regrSxy = unsafeWindowFunctionN "regr_sxy"
-    regrSyy = unsafeWindowFunctionN "regr_syy"
-    stddev = unsafeWindowFunction1 "stddev"
-    stddevPop = unsafeWindowFunction1 "stddev_pop"
-    stddevSamp = unsafeWindowFunction1 "stddev_samp"
-    variance = unsafeWindowFunction1 "variance"
-    varPop = unsafeWindowFunction1 "var_pop"
-    varSamp = unsafeWindowFunction1 "var_samp"
+instance Aggregate (WindowArg grp) (WindowFunction grp) where
+  countStar = UnsafeWindowFunction "count(*)"
+  count = unsafeWindowFunction1 "count"
+  sum_ = unsafeWindowFunction1 "sum"
+  arrayAgg = unsafeWindowFunction1 "array_agg"
+  jsonAgg = unsafeWindowFunction1 "json_agg"
+  jsonbAgg = unsafeWindowFunction1 "jsonb_agg"
+  bitAnd = unsafeWindowFunction1 "bit_and"
+  bitOr = unsafeWindowFunction1 "bit_or"
+  boolAnd = unsafeWindowFunction1 "bool_and"
+  boolOr = unsafeWindowFunction1 "bool_or"
+  every = unsafeWindowFunction1 "every"
+  max_ = unsafeWindowFunction1 "max"
+  min_ = unsafeWindowFunction1 "min"
+  avg = unsafeWindowFunction1 "avg"
+  corr = unsafeWindowFunctionN "corr"
+  covarPop = unsafeWindowFunctionN "covar_pop"
+  covarSamp = unsafeWindowFunctionN "covar_samp"
+  regrAvgX = unsafeWindowFunctionN "regr_avgx"
+  regrAvgY = unsafeWindowFunctionN "regr_avgy"
+  regrCount = unsafeWindowFunctionN "regr_count"
+  regrIntercept = unsafeWindowFunctionN "regr_intercept"
+  regrR2 = unsafeWindowFunctionN "regr_r2"
+  regrSlope = unsafeWindowFunctionN "regr_slope"
+  regrSxx = unsafeWindowFunctionN "regr_sxx"
+  regrSxy = unsafeWindowFunctionN "regr_sxy"
+  regrSyy = unsafeWindowFunctionN "regr_syy"
+  stddev = unsafeWindowFunction1 "stddev"
+  stddevPop = unsafeWindowFunction1 "stddev_pop"
+  stddevSamp = unsafeWindowFunction1 "stddev_samp"
+  variance = unsafeWindowFunction1 "variance"
+  varPop = unsafeWindowFunction1 "var_pop"
+  varSamp = unsafeWindowFunction1 "var_samp"
 
 -- | A `WindowDefinition` is a set of table rows that are somehow related
 -- to the current row
@@ -111,7 +113,7 @@ data WindowDefinition grp lat with db params from where
        -- ^ `Squeal.PostgreSQL.Expression.Sort.orderBy` clause
     -> WindowDefinition grp lat with db params from
 
-instance OrderBy WindowDefinition where
+instance OrderBy (WindowDefinition grp) grp where
   orderBy sortsR (WindowDefinition parts sortsL)
     = WindowDefinition parts (sortsL ++ sortsR)
 
@@ -157,6 +159,43 @@ newtype WindowFunction
     deriving stock (GHC.Generic,Show,Eq,Ord)
     deriving newtype (NFData)
 
+data WindowArg
+  (grp :: Grouping)
+  (args :: [NullType])
+  (lat :: FromType)
+  (with :: FromType)
+  (db :: SchemasType)
+  (params :: [NullType])
+  (from :: FromType)
+    = WindowArg
+    { windowArgs :: NP (Expression grp lat with db params from) args
+    , windowFilter :: [Condition grp lat with db params from]
+    } deriving stock (GHC.Generic)
+
+instance SOP.SListI args
+  => RenderSQL (WindowArg grp args lat with db params from) where
+    renderSQL (WindowArg args filters) =
+      parenthesized (renderCommaSeparated renderSQL args)
+      & renderFilters filters
+      where
+        renderFilter wh = "FILTER" <+> parenthesized ("WHERE" <+> wh)
+        renderFilters = \case
+          [] -> id
+          wh:whs -> (<+> renderFilter (renderSQL (foldr (.&&) wh whs)))
+
+instance FilterWhere (WindowArg grp) grp where
+  filterWhere wh (WindowArg args filters) = WindowArg args (wh : filters)
+
+pattern Window
+  :: Expression grp lat with db params from arg
+  -> WindowArg grp '[arg] lat with db params from
+pattern Window x = Windows (x :* Nil)
+
+pattern Windows
+  :: NP (Expression grp lat with db params from) args
+  -> WindowArg grp args lat with db params from
+pattern Windows xs = WindowArg xs []
+
 instance RenderSQL (WindowFunction grp lat with db params from ty) where
   renderSQL = renderWindowFunction
 
@@ -173,7 +212,7 @@ A @RankNType@ for window functions with 1 argument.
 -}
 type WinFun1 x y
   =  forall grp lat with db params from
-  .  Expression grp lat with db params from x
+  .  WindowArg grp '[x] lat with db params from
      -- ^ input
   -> WindowFunction grp lat with db params from y
      -- ^ output
@@ -184,7 +223,7 @@ Use the `*:` operator to end your argument lists.
 -}
 type WinFunN xs y
   =  forall grp lat with db params from
-  .  NP (Expression grp lat with db params from) xs
+  .  WindowArg grp xs lat with db params from
      -- ^ inputs
   -> WindowFunction grp lat with db params from y
      -- ^ output
@@ -192,12 +231,11 @@ type WinFunN xs y
 -- | escape hatch for defining window functions
 unsafeWindowFunction1 :: ByteString -> WinFun1 x y
 unsafeWindowFunction1 fun x
-  = UnsafeWindowFunction $ fun <> parenthesized (renderSQL x)
+  = UnsafeWindowFunction $ fun <> renderSQL x
 
 -- | escape hatch for defining multi-argument window functions
 unsafeWindowFunctionN :: SOP.SListI xs => ByteString -> WinFunN xs y
-unsafeWindowFunctionN fun xs = UnsafeWindowFunction $ fun <>
-  parenthesized (renderCommaSeparated renderSQL xs)
+unsafeWindowFunctionN fun xs = UnsafeWindowFunction $ fun <> renderSQL xs
 
 {- | rank of the current row with gaps; same as `rowNumber` of its first peer
 
@@ -243,7 +281,7 @@ cumeDist = UnsafeWindowFunction "cume_dist()"
 {- | integer ranging from 1 to the argument value,
 dividing the partition as equally as possible
 
->>> printSQL $ ntile 5
+>>> printSQL $ ntile (Window 5)
 ntile((5 :: int4))
 -}
 ntile :: WinFun1 ('NotNull 'PGint4) ('NotNull 'PGint4)
